People: Move Viewers to separate file
